+++ CHANGELOG.md
@@ -26,12 +26,11 @@ For more information about the _planned features_, please check out our [TODO li
 
 Sometime in one of the future v1.5 patches but before the full new v1.6 minor.
 
-+ pridať BB kód `[img]https://obrazok[/img]` do príspevkov (--)
-
 
 ## Proposed to v1.5.3
 
 * Make the footer links look more technical ([@Kubo2][0])
++ Povolený BB kód `[img]https://obrazok[/img]` do používateľských príspevkov ([@Kubo2][0])
 
 
 ## v1.5.2, release date: 2017-03-30